'Snapshot' Toolbar option opens Snappshots menu
