+++ src/components/Header/ToolbarOptions/Snapshots/SnapshotOption.tsx
@@ -1,27 +1,14 @@
 import { Box } from '@mui/material'
 import Button from '@mui/material/Button'
-import { useState } from 'react'
-import { EditorDB } from '../../../IndexedDB/initDB'
 
 type Props = {
-    onDownload: () => void
-    db: EditorDB
+    openVersions: (e: React.MouseEvent) => void
 }
 
-const FileOption = ({ onDownload, db }: Props) => {
-    const [anchor, setAnchor] = useState<(EventTarget & Element) | null>(null)
-
-    const openMenu = (e: React.MouseEvent) => {
-        setAnchor(e.currentTarget)
-    }
-
-    const closeMenu = () => {
-        setAnchor(null)
-    }
-
+const FileOption = ({ openVersions }: Props) => {
     return (
         <Box sx={{ display: 'inline-block' }}>
-            <Button style={{ padding: 3 }} onClick={openMenu}>
+            <Button style={{ padding: 3 }} onClick={openVersions}>
                 Snapshots
             </Button>
         </Box>
